Ensure AppService Deployment Targets' health check respects proxy settings (#35)
* Ensure HealthCheck respects configured proxy settings
* Extract integration test base class to its own file
* Pull common variable setup down to base class
* Add test to ensure Health Check respects proxy settings
* Ensure Proxy test works on both CI and locally
* Fixes another small difference in testing between CI and local
* More tweaks to test for cross-OS assertion
* Fix up refactor in AppServiceBehaviourFixture
* Simplify Proxy test code
